python-claude-otel-workflow/requirements.txt
claude-agent-sdk>=0.1.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
